✨Add Win Screen and "tutorial"
The win screen gives feedback on player performance. How many correct choices did they make, their score and their overall grade.

The game starts with no timer with just a piece of paper and a recycling bin nearby in order to make it clear what the objective is and to ramp up the complexity over time.